+++ CHANGELOG.md
@@ -1,4 +1,4 @@
 # v2.14.22
 
 * renamed `current_severity_level` to `severity_level` in `MemoryPressureManager`
-
+* deprecated `DictionaryView`